### 2024-05-15 #642 @abigailkeller
eDNAjoint: R package for interpreting paired environmental DNA and traditional surveys
Submitting Author Name: Abigail Keller
Submitting Author Github Handle: @abigailkeller
Repository:  https://github.com/abigailkeller/eDNAjoint
Version submitted: 0.1
Submission type: Stats
Badge grade: silver

Version accepted: TBD
---

```
Package: eDNAjoint
Title: Joint Modeling of Traditional and Environmental DNA Survey Data
Version: 0.1
Maintainer: Abigail G. Keller <agkeller@berkeley.edu>
Author: Abigail G. Keller
    c(person("Abigail G.", "Keller", role = c("aut", "cre"), email="agkeller@berkeley.edu"),
    person("Ryan P.", "Kelly", role = "ctb", email="rpkelly@uw.edu"))
Description: Models integrate environmental DNA (eDNA) detection data and traditional survey data to jointly estimate species catch rate (see package vignette: https://bookdown.org/abigailkeller/eDNAjoint_vignette/). Models can be used with count data via traditional survey methods (i.e., trapping, electrofishing, visual) and replicated eDNA detection/nondetection data via polymerase chain reaction (i.e., PCR or qPCR) from multiple survey locations. Estimated parameters include probability of a false positive eDNA detection, a site-level covariates that scale the sensitivity of eDNA surveys relative to traditional surveys, and catchability coefficients for traditional gear types. Models are implemented with a Bayesian framework (Markov chain Monte Carlo) using the 'Stan' probabilistic programming language.
Roxygen: list(markdown = TRUE, roclets = c ("namespace", "rd", "srr::srr_stats_roclet"))
Biarch: true
    bayestestR,
    loo,
    Rcpp (>= 0.12.0),
    RcppParallel (>= 5.0.1),
    rlist,
    rstan (>= 2.26.23),
    rstantools (>= 2.3.1.1),
    tidyr
    BH (>= 1.66.0),
    Rcpp (>= 0.12.0),
    RcppEigen (>= 0.3.3.3.0),
    RcppParallel (>= 5.0.1),
    rstan (>= 2.26.23),
    StanHeaders (>= 2.26.22)
SystemRequirements: GNU make
    bayesplot,
```

- Please indicate which of our [statistical package categories](https://stats-devguide.ropensci.org/overview.html#overview-categories) this package falls under. (Please check one appropriate box below):

- [x] A pre-submission inquiry has been approved in [issue 628](https://github.com/ropensci/software-review/issues/628)

The package eDNAjoint is useful for interpreting observations from paired environmental DNA (eDNA) and traditional surveys. The package runs a Bayesian model that integrates these two data streams to jointly estimate parameters like the false positive probability of eDNA detection and expected catch rate at a site. The package allows users to access pre-compiled models written in Stan. The target audience is environmental science researchers or managers who want to interpret environmental DNA data but do not have experience writing and implementing custom Bayesian models.

This is the first implementation of a model/algorithm developed in [Keller et al. 2022](https://esajournals.onlinelibrary.wiley.com/doi/full/10.1002/eap.2561).

Silver

Compliance with a good number of standards beyond those identified as minimally necessary.
