### 2024-05-16 #642 @mpadge
Sorry @abigailkeller I'm trying to resolve this. I've worked out that the build fails in building your vignette because of `bayesplot`, but that's clearly not your fault, and nor can I yet work out why. Main problem is that running checks on your package takes hours (on our system at least), so I've thus far only managed to that diagnosis and no further. Once I fix things, you should see a fresh dump of pkgcheck results appear here.
