### 2024-05-17 ropensci/eDNAjoint#2 @abigailkeller
Hi @mpadge Thanks so much for the time you've spent with this!!! So very helpful, and I really appreciate it. A couple thoughts:
1. It definitely makes sense to save some model objects and plots within the package (as .rda files in the data folder?) that can be used to build the vignette. It would be nice to have reviewers be able to build the vignette.
2. An alternative (or in addition?) would be to remove the vignette. I have an [identical copy of the vignette as a bookdown page](https://bookdown.org/abigailkeller/eDNAjoint_vignette/) (also linked in the package's README). I had made this page previously because I find them cleaner to navigate through than the html vignettes.
3. It seems that the size of the src folder should be addressed. I have ideas of how to do this (combining redundant code), but I'm wondering if you have a general ballpark of how much I should be trying to reduce it, especially if I am trying to put the package on Cran?
