### 2024-05-20 #642 @jooolia
Thank you for your submission @abigailkeller ! And thanks @mpadge for looking into the issues running the checks on our system.
Based on the pre-submission issue and the checks (minus the failures that Mark is investigating), I think the only package will be ready to pass on to a handing editor. @abigailkeller do you plan on addressing the duplicated function name "jointModel from JM"?
Thanks, Julia
