+++ tests/testthat/test-jointModel.R
@@ -2,16 +2,16 @@ test_that("jointModel input checks work", {
   #' @srrstats {G5.2,G5.2b,BS2.15} Tests the assure function input checks are
   #'   behaving as expected.
   #1. input tags are valid, q = FALSE, cov = FALSE
-  expect_error(jointModel(data=list(qPCR.n=rbind(c(1,1,1),c(1,1,NA)),
-                                    qPCR.k=rbind(c(3,3,3),c(3,3,NA)),
+  expect_error(jointModel(data=list(qPCR.k=rbind(c(1,1,1),c(1,1,NA)),
+                                    qPCR.n=rbind(c(3,3,3),c(3,3,NA)),
                                     count=rbind(c(4,1,1),c(1,1,NA))),
                           multicore=FALSE),
                "Data should include 'qPCR.N', 'qPCR.K', and 'count'.")
   #2. input tags are valid, q = FALSE, cov = TRUE
   site.cov <- cbind(c(1,0),c(0.4,-0.4))
   colnames(site.cov) <- c('var_a','var_b')
-  expect_error(jointModel(data=list(qPCR.n=rbind(c(1,1,1),c(1,1,NA)),
-                                    qPCR.k=rbind(c(3,3,3),c(3,3,NA)),
+  expect_error(jointModel(data=list(qPCR.k=rbind(c(1,1,1),c(1,1,NA)),
+                                    qPCR.n=rbind(c(3,3,3),c(3,3,NA)),
                                     count=rbind(c(4,1,1),c(1,1,NA)),
                                     site.cov=site.cov),
                           cov=c('var_a','var_b'),
@@ -19,8 +19,8 @@ test_that("jointModel input checks work", {
                paste0("Data should include 'qPCR.N', 'qPCR.K', ",
                       "'count', and 'site.cov'."))
   #3. input tags are valid, q = TRUE, cov = FALSE
-  expect_error(jointModel(data=list(qPCR.n=rbind(c(1,1,1),c(1,1,NA)),
-                                    qPCR.k=rbind(c(3,3,3),c(3,3,NA)),
+  expect_error(jointModel(data=list(qPCR.k=rbind(c(1,1,1),c(1,1,NA)),
+                                    qPCR.n=rbind(c(3,3,3),c(3,3,NA)),
                                     count=rbind(c(4,1,1),c(1,1,NA)),
                                     count.type=rbind(c(1,2,1),c(1,2,NA))),
                           q=TRUE,
@@ -30,8 +30,8 @@ test_that("jointModel input checks work", {
   #4. input tags are valid, q = TRUE, cov = TRUE
   site.cov <- cbind(c(1,0),c(0.4,-0.4))
   colnames(site.cov) <- c('var_a','var_b')
-  expect_error(jointModel(data=list(qPCR.n=rbind(c(1,1,1),c(1,1,NA)),
-                                    qPCR.k=rbind(c(3,3,3),c(3,3,NA)),
+  expect_error(jointModel(data=list(qPCR.k=rbind(c(1,1,1),c(1,1,NA)),
+                                    qPCR.n=rbind(c(3,3,3),c(3,3,NA)),
                                     count=rbind(c(4,1,1),c(1,1,NA)),
                                     count.type=rbind(c(1,2,1),c(1,2,NA)),
                                     site.cov=site.cov),
@@ -43,8 +43,8 @@ test_that("jointModel input checks work", {
   #5. make sure dimensions of qPCR.N and qPCR.K are equal
   #' @srrstats {BS2.1a} Test to ensure pre-processing routines to ensure all
   #'   input data is dimensionally commensurate
[…]
