+++ R/srr-stats-standards.R
@@ -100,6 +100,27 @@ NULL
 #'   sampling, and using something like Sys.time() doesn't seem like it would
 #'   provide an accurate estimate of computation time, but I'm open to other
 #'   suggestions.
+#' @srrstatsNA {BS1.2b, BS1.2c} Vignette was removed from package because of
+#'   build size, but textual descriptions of priors and examples can be found in
+#'   sections 4.3 and 2.5 in the user guide
+#'   (https://bookdown.org/abigailkeller/eDNAjoint_vignette/).
+#' @srrstatsNA {G1.5} Vignette was removed from package because of
+#'   build size, but example code that reproduces results in the publication
+#'   (Keller et al., 2022) where the model/algorithm was first developed can be
+#'   found in sections 2.2 and 3.2 in the user guide
+#'   (https://bookdown.org/abigailkeller/eDNAjoint_vignette/).
+#' @srrstatsNA {BS6.3} Vignette was removed from package because of
+#'   build size, but examples of plotting posterior distributional estimates can
+#'   be found in sections 2.4.1 and 3.4.1 in the user guide
+#'   (https://bookdown.org/abigailkeller/eDNAjoint_vignette/).
+#' @srrstatsNA {BS1.4,BS4.3,BS6.2} Vignette was removed from package because of
+#'   build size, but examples of convergence checking and plotting posterior
+#'   samples can be found in sections 2.4.1 and 3.4.1 in the user guide
+#'   (https://bookdown.org/abigailkeller/eDNAjoint_vignette/).
+#' @srrstatsNA {BS4.0} Vignette was removed from package because of
+#'   build size, but documentation of sampling algorithm can be found in
+#'   section 4.2 in the user guide
+#'   (https://bookdown.org/abigailkeller/eDNAjoint_vignette/).
 #'
 #'
 #' Any non-applicable standards can have their tags changed from `@srrstatsTODO`
