### 2024-05-26 #642 @abigailkeller
Update: since my vignette compilation time was causing build failures ([see issue here](https://github.com/abigailkeller/eDNAjoint/issues/2#issuecomment-2128848886)), I removed the vignette from the package and instead will direct users/reviewers to a copy of my vignette outside the package in [this book](https://bookdown.org/abigailkeller/eDNAjoint_vignette/). This book is linked in the package DESCRIPTION, as well as in the documentation in all functions.
