### 2024-05-26 ropensci/eDNAjoint#2 @abigailkeller
Hi @mpadge thanks again for your help with this. A few responses:

1. I just pushed a version of the package without the vignette and linked the [user guide](https://bookdown.org/abigailkeller/eDNAjoint_vignette/) in the package DESCRIPTION and in the documentation for all functions. I can still try and reduce the `/src` directory if needed!

2. Thanks for the idea of the probability distribution standards. First just to clarify how the package is structured (and this makes me think I should do a better job communicating the package): the package stores a bunch of Bayesian models written in Stan [in this folder](https://github.com/abigailkeller/eDNAjoint/tree/master/inst/stan). Each file here contains a variation of a model that uses environmental DNA data and "traditional" survey data to jointly estimate parameters. These model variations are accessed based on the type of input data and/or user-defined input parameters, including distributional assumptions. And then rstantools is used to compile the models that you find in the `/src` folder. These compiled models are then called during MCMC sampling.

3. I think the following probability distribution standards may apply:

- PD1.0: Some of the distributions in the models are user-defined and others are fixed and assumed. I can definitely provide documentation for why these were chosen, but I'm curious where this documentation should go? The natural location for these might be in the .Stan files themselves where the models are defined ([example here](https://github.com/abigailkeller/eDNAjoint/blob/master/inst/stan/joint_binary_cov_negbin.stan)), but I'm not quite sure the `srr` syntax would work in a non-R file.
- PD4.0: In my tests, I do "parameter recovery" tests, i.e., simulate data with known parameter values, run a package function with the data, and then test if the known parameter values are within the 95% credibility interval of the parameters' posteriors in the function output. This seems like how one would test for numeric equality in a Bayesian context? Although curious if you have other advice.
- PD4.1: I'm not sure this makes sense here because the output values of MCMC are derived stochastically. I use the MCMC algorithm in rstan's `sampling` function, rather than my own MCMC algorithm, so I'm not sure I would be able to reproduce how the output values are derived.
- PD4.2: This standard definitely makes sense for this package. One input parameter allows the user to make a distributional assumption about the traditional survey data. I can definitely add a test that shows that results will differ based on this assumption.
