### 2024-05-27 ropensci/eDNAjoint#2 @mpadge
That sounds great, and sounds to me like it would be useful for you to comply with prob. dist. standards. The abiding aim of our whole standardisation procedures is to improve software quality, and so if you think that compliance with these extra standards will in any way improve either your software itself, or the ability of others to understand why your software is the way it is, then please extend you compliance to include these extra standards.

Specific thoughts in response to your comments:

>  1. I just pushed a version of the package without the vignette and linked the [user guide](https://bookdown.org/abigailkeller/eDNAjoint_vignette/) in the package DESCRIPTION and in the documentation for all functions. I can still try and reduce the `/src` directory if needed!

No need, that should be fine now.

>  2. Thanks for the idea of the probability distribution standards. First just to clarify how the package is structured (and this makes me think I should do a better job communicating the package): the package stores a bunch of Bayesian models written in Stan [in this folder](https://github.com/abigailkeller/eDNAjoint/tree/master/inst/stan?rgh-link-date=2024-05-26T15%3A29%3A57Z). Each file here contains a variation of a model that uses environmental DNA data and "traditional" survey data to jointly estimate parameters. These model variations are accessed based on the type of input data and/or user-defined input parameters, including distributional assumptions. And then rstantools is used to compile the models that you find in the `/src` folder. These compiled models are then called during MCMC sampling.

One suggestion for where to document this would be in a `README.md` file in the `/src` folder itself. GitHub will directly display README files in any sub-directory, and this is a really useful way to document sub-directory specific stuff like that. I'm not sure the `srr` package will currently correctly parse that, but feel free to open an issue [there](https://github.com/ropensci-review-tools/srr) and we can easily accommodate that.

>  * PD1.0: Some of the distributions in the models are user-defined and others are fixed and assumed. I can definitely provide documentation for why these were chosen, but I'm curious where this documentation should go? The natural location for these might be in the .Stan files themselves where the models are defined ([example here](https://github.com/abigailkeller/eDNAjoint/blob/master/inst/stan/joint_binary_cov_negbin.stan?rgh-link-date=2024-05-26T15%3A29%3A57Z)), but I'm not quite sure the `srr` syntax would work in a non-R file.

Yes, it works in any files in the `/src` directory too. As above, open an issue in `srr` if you encounter any problems there.

>  * PD4.0: In my tests, I do "parameter recovery" tests, i.e., simulate data with known parameter values, run a package function with the data, and then test if the known parameter values are within the 95% credibility interval of the parameters' posteriors in the function output. This seems like how one would test for numeric equality in a Bayesian context? Although curious if you have other advice.

Sounds great, and definitely sounds like it would be useful to document that.

>  * PD4.1: I'm not sure this makes sense here because the output values of MCMC are derived stochastically. I use the MCMC algorithm in rstan's `sampling` function, rather than my own MCMC algorithm, so I'm not sure I would be able to reproduce how the output values are derived.

Documenting that, including statements of inability to strictly comply, would then be all the more important.

>  * PD4.2: This standard definitely makes sense for this package. One input parameter allows the user to make a distributional assumption about the traditional survey data. I can definitely add a test that shows that results will differ based on this assumption.

That also sounds like it would be particularly helpful to document.

Thanks for your considered responses!

---

Also note that your submission made me realise that we had neglected to update our submission templates to include probability distribution standards. If you do decide to also comply with those standards, could you please edit your [submission template](https://github.com/ropensci/software-review/issues/642#issue-2296785284) and include an additional checked item for "Probability Distributions". Thanks!
