+++ src/stanExports_joint_binary_gamma.h
@@ -29,62 +29,62 @@ using namespace stan::math;
 stan::math::profile_map profiles__;
 static constexpr std::array<const char*, 57> locations_array__ =
   {" (found before start of program)",
-  " (in 'joint_binary_gamma', line 13, column 4 to column 42)",
-  " (in 'joint_binary_gamma', line 14, column 4 to column 45)",
-  " (in 'joint_binary_gamma', line 15, column 4 to column 23)",
-  " (in 'joint_binary_gamma', line 16, column 4 to column 26)",
-  " (in 'joint_binary_gamma', line 19, column 2 to column 31)",
-  " (in 'joint_binary_gamma', line 20, column 2 to column 43)",
-  " (in 'joint_binary_gamma', line 21, column 2 to column 40)",
-  " (in 'joint_binary_gamma', line 22, column 2 to column 33)",
-  " (in 'joint_binary_gamma', line 46, column 2 to column 22)",
-  " (in 'joint_binary_gamma', line 47, column 2 to column 11)",
-  " (in 'joint_binary_gamma', line 24, column 4 to column 41)",
-  " (in 'joint_binary_gamma', line 25, column 4 to column 41)",
-  " (in 'joint_binary_gamma', line 26, column 4 to column 33)",
-  " (in 'joint_binary_gamma', line 23, column 19 to line 27, column 3)",
-  " (in 'joint_binary_gamma', line 23, column 2 to line 27, column 3)",
-  " (in 'joint_binary_gamma', line 29, column 6 to column 42)",
-  " (in 'joint_binary_gamma', line 28, column 15 to line 30, column 5)",
-  " (in 'joint_binary_gamma', line 28, column 2 to line 30, column 5)",
-  " (in 'joint_binary_gamma', line 48, column 2 to column 21)",
-  " (in 'joint_binary_gamma', line 50, column 10 to column 84)",
-  " (in 'joint_binary_gamma', line 49, column 17 to line 51, column 7)",
-  " (in 'joint_binary_gamma', line 49, column 4 to line 51, column 7)",
-  " (in 'joint_binary_gamma', line 53, column 10 to column 61)",
-  " (in 'joint_binary_gamma', line 52, column 19 to line 54, column 7)",
-  " (in 'joint_binary_gamma', line 52, column 6 to line 54, column 7)",
-  " (in 'joint_binary_gamma', line 34, column 6 to column 61)",
-  " (in 'joint_binary_gamma', line 33, column 17 to line 35, column 5)",
-  " (in 'joint_binary_gamma', line 33, column 4 to line 35, column 5)",
-  " (in 'joint_binary_gamma', line 37, column 8 to column 39)",
-  " (in 'joint_binary_gamma', line 36, column 18 to line 38, column 5)",
-  " (in 'joint_binary_gamma', line 36, column 4 to line 38, column 5)",
-  " (in 'joint_binary_gamma', line 40, column 2 to column 47)",
-  " (in 'joint_binary_gamma', line 41, column 2 to column 22)",
-  " (in 'joint_binary_gamma', line 42, column 2 to column 32)",
-  " (in 'joint_binary_gamma', line 43, column 2 to column 33)",
-  " (in 'joint_binary_gamma', line 2, column 4 to column 19)",
-  " (in 'joint_binary_gamma', line 3, column 4 to column 19)",
-  " (in 'joint_binary_gamma', line 4, column 10 to column 11)",
-  " (in 'joint_binary_gamma', line 4, column 4 to column 28)",
-  " (in 'joint_binary_gamma', line 5, column 10 to column 11)",
-  " (in 'joint_binary_gamma', line 5, column 4 to column 28)",
-  " (in 'joint_binary_gamma', line 6, column 4 to column 22)",
[…]
