+++ src/stanExports_joint_binary_negbin.h
@@ -29,53 +29,53 @@ using namespace stan::math;
 stan::math::profile_map profiles__;
 static constexpr std::array<const char*, 48> locations_array__ =
   {" (found before start of program)",
-  " (in 'joint_binary_negbin', line 14, column 4 to column 33)",
-  " (in 'joint_binary_negbin', line 15, column 4 to column 22)",
-  " (in 'joint_binary_negbin', line 16, column 4 to column 23)",
-  " (in 'joint_binary_negbin', line 17, column 4 to column 26)",
-  " (in 'joint_binary_negbin', line 20, column 2 to column 43)",
-  " (in 'joint_binary_negbin', line 21, column 2 to column 40)",
-  " (in 'joint_binary_negbin', line 40, column 2 to column 22)",
-  " (in 'joint_binary_negbin', line 41, column 2 to column 11)",
-  " (in 'joint_binary_negbin', line 23, column 4 to column 41)",
-  " (in 'joint_binary_negbin', line 24, column 4 to column 33)",
-  " (in 'joint_binary_negbin', line 22, column 19 to line 25, column 3)",
-  " (in 'joint_binary_negbin', line 22, column 2 to line 25, column 3)",
-  " (in 'joint_binary_negbin', line 42, column 2 to column 21)",
-  " (in 'joint_binary_negbin', line 44, column 10 to column 65)",
-  " (in 'joint_binary_negbin', line 43, column 17 to line 45, column 7)",
-  " (in 'joint_binary_negbin', line 43, column 4 to line 45, column 7)",
-  " (in 'joint_binary_negbin', line 47, column 10 to column 61)",
-  " (in 'joint_binary_negbin', line 46, column 19 to line 48, column 7)",
-  " (in 'joint_binary_negbin', line 46, column 6 to line 48, column 7)",
-  " (in 'joint_binary_negbin', line 29, column 8 to column 45)",
-  " (in 'joint_binary_negbin', line 28, column 18 to line 30, column 5)",
-  " (in 'joint_binary_negbin', line 28, column 4 to line 30, column 5)",
-  " (in 'joint_binary_negbin', line 32, column 8 to column 39)",
-  " (in 'joint_binary_negbin', line 31, column 18 to line 33, column 5)",
-  " (in 'joint_binary_negbin', line 31, column 4 to line 33, column 5)",
-  " (in 'joint_binary_negbin', line 35, column 2 to column 47)",
-  " (in 'joint_binary_negbin', line 36, column 2 to column 22)",
-  " (in 'joint_binary_negbin', line 37, column 2 to column 42)",
-  " (in 'joint_binary_negbin', line 2, column 4 to column 19)",
-  " (in 'joint_binary_negbin', line 3, column 4 to column 19)",
-  " (in 'joint_binary_negbin', line 4, column 10 to column 11)",
-  " (in 'joint_binary_negbin', line 4, column 4 to column 28)",
-  " (in 'joint_binary_negbin', line 5, column 10 to column 11)",
-  " (in 'joint_binary_negbin', line 5, column 4 to column 28)",
-  " (in 'joint_binary_negbin', line 6, column 4 to column 22)",
-  " (in 'joint_binary_negbin', line 7, column 10 to column 11)",
-  " (in 'joint_binary_negbin', line 7, column 4 to column 28)",
-  " (in 'joint_binary_negbin', line 8, column 10 to column 11)",
-  " (in 'joint_binary_negbin', line 8, column 4 to column 28)",
-  " (in 'joint_binary_negbin', line 9, column 10 to column 11)",
-  " (in 'joint_binary_negbin', line 9, column 4 to column 28)",
-  " (in 'joint_binary_negbin', line 10, column 4 to column 28)",
[…]
