### 2024-05-30 ropensci/eDNAjoint#2 @abigailkeller
@mpadge Ok, great! Thanks so much for your help, again.

I just pushed a version of the package with the probability distribution standards included. I added some standards to a .Rmd file (that generates a github_document), but it looks like `srr` does not recognize the inclusion of the standards in this way. I'll create an issue with that package.

Thanks!!
