+++ README.md
@@ -9,6 +9,8 @@
 
 [![R-CMD-check](https://github.com/abigailkeller/eDNAjoint/actions/workflows/R-CMD-check.yaml/badge.svg)](https://github.com/abigailkeller/eDNAjoint/actions/workflows/R-CMD-check.yaml)
 [![codecov](https://codecov.io/gh/abigailkeller/eDNAjoint/graph/badge.svg?token=AEVR9NSQ9Z)](https://codecov.io/gh/abigailkeller/eDNAjoint)
+[![Status at rOpenSci Software Peer
+Review](https://badges.ropensci.org/642_status.svg)](https://github.com/ropensci/software-review/issues/642)
 <!-- badges: end -->
 
 ``` r
@@ -66,7 +68,7 @@ detection, $p_{10}$:
 # summarize p10 posterior
 jointSummarize(goby.fit$model, par = 'p10')
 #>      mean se_mean    sd  2.5% 97.5%    n_eff Rhat
-#> p10 0.003       0 0.001 0.001 0.007 18928.52    1
+#> p10 0.003       0 0.001 0.001 0.007 17656.22    1
 ```
 
 Or to find the number of eDNA samples and traditional survey samples
