+++ NEWS.md
@@ -1,0 +1,3 @@
+# eDNAjoint 0.1
+
+-   First version of eDNAjoint
