### 2024-06-27 #642 @abigailkeller
@emitanaka Hi! I'm wondering if it is too late in the review process to express interest in publishing in Methods in Ecology and Evolution? I am thinking that publishing in this journal may make the package more visible to my intended audience. However, I know that this was something I should have indicated when creating this issue, so I understand if it is not possible!

Thanks,
Abby
