### 2024-07-17 #642 @emitanaka
@abigailkeller re: submitting to MEE, I think it should not be a problem. You may want to complete the review here first and quote that the package has been reviewed by ropensci. This fastened the case for [JOSS](https://discuss.ropensci.org/t/automated-joss-review-following-ropensci-acceptance/2807).
