### 2024-07-19 #642 @smwindecker
-   [x] As the reviewer I confirm that there are no [conflicts of interest](https://devguide.ropensci.org/policies.html#coi) for me to review this work (If you are unsure whether you are in conflict, please speak to your editor *before* starting your review).

### Compliance with Standards

-   [x] This package complies with a sufficient number of standards for a silver badge
-   [x] This grade of badge is the same as what the authors wanted to achieve

The following standards currently deemed non-applicable (through tags of `@srrstatsNA`) could potentially be applied to future versions of this software: (Please specify)

> G1.2: The reason this was erroring was because the function is looking for "list(markdown = TRUE)" in the description file. The additional content you have ", roclets = c ("namespace", "rd", "srr::srr_stats_roclet"))" is making the uses_markdown_md() return an NA value and thereby erroring the use_lifecycle() function, as described in <https://rdrr.io/cran/usethis/src/R/roxygen.R#sym-uses_roxygen_md>. If you temporarily remove the roclets argument, you can run the `usethis::use_lifecycle()` then paste back the roclets argument into the description file afterwards.

> BS6.0/BS6.1 re default print and plot methods. We agree with the authors that the existing packages that exist to plot stan outputs are sufficient and they do not need to provide their own. I also think the demonstrations and examples in the bookdown are very helpful. We also agree that the rstan.fit object has its own print function, however the jointModel output contains not only the rstan.fit object but also a list of the inits. If they would like, the authors could perhaps try to return the inits in a neater format, for the moment they are returned in an unnamed list of 4, which, to a novice user, may not be immediately obvious correpond to the 4 chains.

Please also comment on any standards which you consider particularly well documented.

> G1.0 This software cites as a primary reference a paper written by the authors making use of this model in an applied context. This is an excellent reference because it demonstrates that the method has already been peer-reviewed.

> G1.3 The bookdown documentation is very thorough, and walks through not only the relevant statistical terminology very clearly, but creates helpful ties to the relevant data use case.

> G5.1 the gobyData is clear, easy to use, and well documented in the bookdown. It is easy to test out the functionality with this data.

> BS1.1 I think the documentation in the bookdown related to how to transform/reshape input data is especially well written. It gives users a great guide for converting their data into the right format.

For packages aiming for silver or gold badges:

-   [x] This package extends beyond minimal compliance with standards in the following ways: (please describe)

> It is our assessment that the package complies with more than half of the potentially applicable standards.

### General Review

-   [ ] **Community guidelines** including contribution guidelines in the README or CONTRIBUTING
-   [x] The documentation is sufficient to enable general use of the package beyond one specific use case

> Some guidelines for contributing could be listed explicitly if the authors would welcome them.

#### Algorithms

-   How well are algorithms encoded?
-   Is the choice of computer language appropriate for that algorithm, and/or envisioned use of package?
-   Are aspects of algorithmic scaling sufficiently documented and tested?
-   Are there any aspects of algorithmic implementation which could be improved?

> The wrapper to build the stan models is very well designed. It allows users the option to specify a gamma distribution for the overdispersion of survey data and a beta distribution for the eDNA false positive rate. These options are well selected for the subject matter and provide a great resource for researchers in this area.

-   Regardless of actual coverage of tests, are there any fundamental software operations which are not sufficiently expressed in tests?
-   Is there a need for extended tests, or if extended tests exists, have they been implemented in an appropriate way, and are they appropriately documented?

> There is a good set of unit tests provided with the package. Improperly formatted input data gives useful errors to revise input.

#### Visualisation (where appropriate)

-   Do visualisations aid the primary purposes of statistical interpretation of results?
-   Are there any aspects of visualisations which could risk statistical misinterpretation?

> In general the existing visualisations in other packages for stan model outputs are sufficient for visualisation of the outputs from this package. The guidelines in the bookdown are also very clear. However for users that don't work through the bookdown itself, if the authors were interested, they could perhaps point to these options in the help file for the model fit function or in a print message for the fitted object. I spoke to a few colleagues who use similar packages that wrap stan and they agreed they wouldn't expect bespoke plotting functionality.

#### Package Design

-   Is the package well designed for its intended purpose?
-   In relation to **External Design:** Do exported functions and the relationships between them enable general usage of the package?
-   In relation to **External Design:** Do exported functions best serve inter-operability with other packages?
-   In relation to **Internal Design:** Are algorithms implemented appropriately in terms of aspects such as efficiency, flexibility, generality, and accuracy?
-   In relation to **Internal Design:** Could ranges of admissible input structures, or form(s) of output structures, be expanded to enhance inter-operability with other packages?

> The package is well designed. It provides a modelling solution for paired eDNA and traditional sampling data. By building an integrated model with this paired data, it enables users to compare the two survey protocols and get better estimates for the number of the target species.

> As an aside, after speaking with a colleague who works in eDNA, they expressed interest in the intent of the package but noted that non perfectly paired data is more common. It would be great in the future if the authors could consider an extension that would allow users to input non-paired data, to allow more potential users to get the most out of available datasets (though I understand this is a much thornier model, just thought I'd pass along the note).

> In terms of althorithmic implementation the authors have done a good job designing the model in stan. The jointModel function is nicely written to include wrapper arguments for the MCMC options and easy input location for initial values.

> I particularly like the demonstration in the bookdown for simulating from the prior, to help inform model development.

Reviewer 1, 9 hours, reviewer 2, 12 hours

-   [x] Should the author(s) deem it appropriate, we agree to be acknowledged as a package reviewer ("rev" role) in the package DESCRIPTION file.

> I like that the helpfiles link to the bookdown.

> Regarding BS1.3a, BS2.8, and running the model with the results of previous runs: I was unsure what the standards were suggesting, but if they were suggesting using posteriors as the starting point for continuing the Markov chains, I agree with the authors' assessment that they are not sure how to implement this. I found this discussion that the authors may find interesting on this point: <https://discourse.mc-stan.org/t/start-from-previous-simulation/8282/6>

> We had one issue with installation, which was that for one of us the install took around 30 minutes to complete. This seems unusally long. If this install time cannot be reduced, perhaps just a note in the readme warning users that it may take a long time. The clang messages that come along with the compiled stan models are confronting for a user that has not encountered them before, and may make someone interrupt installation if it is taking so long.

> A few notes from some automatic checks:

> -   The following function have high cyclomatic complexity (\>50): detectionCalculate (117), detectionPlot (117), jointSummarize (105), jointModel (66). You can make them easier to reason about by encapsulating distinct steps of your function into subfunctions.

> -   use '\<-' for assignment instead of '='. '\<-' is the standard, and R users and developers are used it and it is easier to read your code for them if you use '\<-'. See in tests/testthat/test-jointSelect.R:35:10 tests/testthat/test-jointSelect.R:38:10 tests/testthat/test-muCritical.R:57:11 tests/testthat/test-muCritical.R:63:9

> -   style consistency: add spaces around "=" for better readability.
