### 2024-07-25 5cbfe12
reduced cyclomatic complexity
