+++ R/detectionCalculate.R
@@ -51,17 +51,17 @@
 #' data(gobyData)
 #'
 #' # Fit a model including 'Filter_time' and 'Salinity' site-level covariates
-#' fit.cov = jointModel(data=gobyData, cov=c('Filter_time','Salinity'),
-#'                      family="poisson", p10priors=c(1,20), q=FALSE,
-#'                      multicore=FALSE)
+#' fit.cov <- jointModel(data = gobyData, cov = c('Filter_time','Salinity'),
+#'                       family = "poisson", p10priors = c(1,20), q = FALSE,
+#'                       multicore = FALSE)
 #'
 #' # Calculate at the mean covariate values
-#' # (covariates are standardized, so mean=0)
-#' detectionCalculate(fit.cov$model, mu = seq(from=0.1,to=1,by=0.1),
+#' # (covariates are standardized, so mean = 0)
+#' detectionCalculate(fit.cov$model, mu = seq(from = 0.1, to = 1, by = 0.1),
 #'                    cov.val = c(0,0), qPCR.N = 3)
 #'
 #' # Calculate mu_critical at salinity 0.5 z-scores greater than the mean
-#' detectionCalculate(fit.cov$model, mu = seq(from=0.1,to=1,by=0.1),
+#' detectionCalculate(fit.cov$model, mu = seq(from = 0.1, to = 1, by = 0.1),
 #'                    cov.val = c(0,0.5), qPCR.N = 3)
 #'
 #' # Ex. 2: Calculating necessary effort for detection with multiple traditional
@@ -71,11 +71,11 @@
 #' data(greencrabData)
 #'
 #' # Fit a model with no site-level covariates
-#' fit.q = jointModel(data=greencrabData, cov=NULL, family="negbin",
-#'                    p10priors=c(1,20), q=TRUE, multicore=FALSE)
+#' fit.q <- jointModel(data = greencrabData, cov = NULL, family = "negbin",
+#'                     p10priors = c(1,20), q = TRUE, multicore = FALSE)
 #'
 #' # Calculate
-#' detectionCalculate(fit.q$model, mu = seq(from=0.1,to=1,by=0.1),
+#' detectionCalculate(fit.q$model, mu = seq(from = 0.1, to = 1, by = 0.1),
 #'                    cov.val = NULL, qPCR.N = 3)
 #'
 #' # Change probability of detecting presence to 0.95
@@ -84,7 +84,7 @@
 #' }
 #'
 
-detectionCalculate <- function(modelfit, mu, cov.val = NULL, probability=0.9,
+detectionCalculate <- function(modelfit, mu, cov.val = NULL, probability = 0.9,
                                qPCR.N = 3){
 
   # input checks
@@ -99,10 +99,10 @@ detectionCalculate <- function(modelfit, mu, cov.val = NULL, probability=0.9,
   ## check to see if there are any divergent transitions
   #' @srrstats {BS4.5} Warning message if the input model fit has divergence
   #'   transitions
-  if(sum(lapply(rstan::get_sampler_params(modelfit,inc_warmup=FALSE),
+  if(sum(lapply(rstan::get_sampler_params(modelfit,inc_warmup = FALSE),
                 div_check)[[1]]) > 0){
 
-    sum <- sum(lapply(rstan::get_sampler_params(modelfit,inc_warmup=FALSE),
+    sum <- sum(lapply(rstan::get_sampler_params(modelfit,inc_warmup = FALSE),
                       div_check)[[1]])
 
     warning <- paste0('Warning: There are ',sum,
@@ -111,619 +111,171 @@ detectionCalculate <- function(modelfit, mu, cov.val = NULL, probability=0.9,
 
   }
 
[…]
