### 2024-07-28 be9cd06
fixed bugs with cyclomatic complexity update
