+++ man/jointSummarize.Rd
@@ -47,8 +47,8 @@ If any of these checks fail, the function returns an error message.
 data(greencrabData)
 
 # Fit a model
-modelfit = jointModel(data=greencrabData, family="negbin", q=TRUE,
-                      multicore=FALSE)
+modelfit <- jointModel(data = greencrabData, family = "negbin", q = TRUE,
+                       multicore = FALSE)
 
 # Create summary table of all parameters
 jointSummarize(modelfit$model)
