+++ tests/testthat/test-jointModel.R
@@ -930,31 +930,29 @@ test_that("jointModel parameter recovery tests work",{
     site.cov = mat_site
   )
   # run model
-  fit_large <- suppressWarnings({jointModel(data = data, 
+  fit_large <- suppressWarnings({jointModel(data = data,
                                             cov = c('var_a','var_b'),
                                             multicore = FALSE, seed = 10)})
   # summary
   summary_large <- as.data.frame(rstan::summary(fit_large$model,
                                                 pars = 'alpha',
-                                                probs = c(0.025, 
+                                                probs = c(0.025,
                                                           0.975))$summary)
 
-  # compare standard errors of estimates of alpha with implementation
-  # with a smaller dataset
-  se_large <- c(summary_large['alpha[1]','se_mean'],
-                summary_large['alpha[2]','se_mean'],
+  # compare standard errors of estimates of alpha coefficients with
+  # implementation with a smaller dataset
+  se_large <- c(summary_large['alpha[2]','se_mean'],
                 summary_large['alpha[3]','se_mean']
   )
 
   # get values for smaller dataset
-  se_small <- c(summary1['alpha[1]','se_mean'],
-                summary1['alpha[2]','se_mean'],
+  se_small <- c(summary1['alpha[2]','se_mean'],
                 summary1['alpha[3]','se_mean']
   )
 
   # set up empty vector to check if standard errors are smaller with
   # larger dataset
-  check <- rep(NA,length(alpha))
+  check <- rep(NA,length(alpha)-1)
   for(i in seq_along(check)){
     check[i] <- se_large[i] < se_small[i]
   }
@@ -963,7 +961,7 @@ test_that("jointModel parameter recovery tests work",{
   #'   properties of data change (i.e., standard error of posteriors is smaller
   #'   if there are more data observations)
   # all should be equal to true
-  expect_equal(check,rep(TRUE,length(alpha)))
+  expect_equal(check,rep(TRUE,length(alpha)-1))
 
 })
 
@@ -1055,3 +1053,1020 @@ test_that("jointModel probability distribution tests work",{
 
 
 })
+
+test_that("semi-paired model works", {
+
+  ## 1.
+  # model includes 'p10','q','phi','beta'
+
+  # constants
+  nsite <- 20
+  nobs_count <- 100
+  nobs_pcr <- 8
+  # params
+  mu <- rlnorm(nsite, meanlog = log(1), sdlog = 1)
+  beta <- 0.5
+  log_p10 <- -4.5
+  q <- 2
+  phi <- 1.2
+  # traditional type
+  count_type <- cbind(matrix(1, nrow = nsite, ncol = nobs_count/2),
+                      matrix(2, nrow = nsite, ncol = nobs_count/2))
+
+  # count
+  count <- matrix(NA, nrow = nsite, ncol = nobs_count)
+  for(i in 1:nsite){
+    for(j in 1:nobs_count){
+      if(count_type[i,j]==1){
+        count[i,j] <- rnbinom(n = 1, mu = mu[i], size = phi)
+      } else {
[…]
