### 2024-08-05 cda116d
bug in jointModel and contributing docs
