+++ README.Rmd
@@ -1,5 +1,8 @@
 ---
 output: github_document
+editor_options: 
+  markdown: 
+    wrap: 72
 ---
 
 <!-- README.md is generated from README.Rmd. Please edit that file -->
@@ -14,25 +17,41 @@ knitr::opts_chunk$set(
 ```
 
 # eDNAjoint
-<img src="man/figures/logo.png" align="right" height="200" dpi="700" />
+
+<img src="man/figures/logo.png" align="right" height="200" dpi="700"/>
 
 <!-- badges: start -->
+
 [![R-CMD-check](https://github.com/abigailkeller/eDNAjoint/actions/workflows/R-CMD-check.yaml/badge.svg)](https://github.com/abigailkeller/eDNAjoint/actions/workflows/R-CMD-check.yaml)
 [![codecov](https://codecov.io/gh/abigailkeller/eDNAjoint/graph/badge.svg?token=AEVR9NSQ9Z)](https://codecov.io/gh/abigailkeller/eDNAjoint)
-[![Status at rOpenSci Software Peer Review](https://badges.ropensci.org/642_status.svg)](https://github.com/ropensci/software-review/issues/642)
-<!-- badges: end -->
+[![Status at rOpenSci Software Peer
+Review](https://badges.ropensci.org/642_status.svg)](https://github.com/ropensci/software-review/issues/642)
+[![Lifecycle:
+experimental](https://img.shields.io/badge/lifecycle-experimental-orange.svg)](https://lifecycle.r-lib.org/articles/stages.html#experimental)
 
+<!-- badges: end -->
 
 ```{r, include = FALSE}
 #' @srrstats {BS1.2a} README with high-level overview and examples of specifying
 #'   prior distributions here
 ```
 
-The package *eDNAjoint* is useful for interpreting observations from paired environmental DNA (eDNA) and traditional surveys. The package runs a Bayesian model that integrates these two data streams to jointly estimate parameters like the false positive probability of eDNA detection and expected catch rate at a site. Optional model variations allow inclusion of site-level covariates that scale the sensitivity of eDNA sampling relative to traditional sampling, as well as estimation of catchability coefficients when multiple traditional gear types are used. Additional functions in the package facilitate interpretation of model fits.
+The package *eDNAjoint* is useful for interpreting observations from
+paired or semi-paired environmental DNA (eDNA) and traditional surveys.
+The package runs a Bayesian model that integrates these two data streams
+to jointly estimate parameters like the false positive probability of
+eDNA detection and expected catch rate at a site. Optional model
+variations allow inclusion of site-level covariates that scale the
+sensitivity of eDNA sampling relative to traditional sampling, as well
+as estimation of catchability coefficients when multiple traditional
+gear types are used. Additional functions in the package facilitate
+interpretation of model fits.
 
 ## Installation
 
-You can install the development version of eDNAjoint:
+You can install the development version of eDNAjoint. Note that
+installation time from source can be high, as unlike binary file formats
+provided by Cran, the program needs to compile during installation:
 
 ```{r,eval=FALSE}
 library(devtools)
@@ -41,9 +60,17 @@ install_github('abigailkeller/eDNAjoint')
 
 ## Example
 
-The main functionality in *eDNAjoint* is the use of `jointModel()` that will fit the model to data. Further functions like `jointSummarize()` and `detectionCalculate()` can be used to help with model fit interpretation.
+The main functionality in *eDNAjoint* is the use of `jointModel()` that
+will fit the model to data. Further functions like `jointSummarize()`
+and `detectionCalculate()` can be used to help with model fit
+interpretation.
 
[…]
