+++ DESCRIPTION
@@ -20,11 +20,12 @@ Imports:
     bayestestR,
     dplyr,
     ggplot2,
+    lifecycle,
     loo,
     magrittr,
     methods,
     Rcpp (>= 0.12.0),
-    RcppParallel (>= 5.0.1), 
+    RcppParallel (>= 5.0.1),
     rlist,
     rstan (>= 2.26.23),
     rstantools (>= 2.3.1.1),

[…]
