### 2024-08-06 839a626
fixed typos in contributing docs
