+++ tests/testthat/test-jointModel.R
@@ -6,7 +6,11 @@ test_that("jointModel input checks work", {
                                       qPCR.n = rbind(c(3,3,3),c(3,3,NA)),
                                       count = rbind(c(4,1,1),c(1,1,NA))),
                           multicore = FALSE),
-               "Data should include 'qPCR.N', 'qPCR.K', and 'count'.")
+               paste("Data should include 'qPCR.N', 'qPCR.K', and 'count'.",
+                      'See the eDNAjoint guide for data formatting help: ',
+                      paste0('https://bookdown.org/abigailkeller/eDNAjoint_',
+                             'vignette/usecase1.html#prepare-the-data'),
+                     sep='\n'))
   #2. input tags are valid, q = FALSE, cov = TRUE
   site.cov <- cbind(c(1,0),c(0.4,-0.4))
   colnames(site.cov) <- c('var_a','var_b')
@@ -16,8 +20,12 @@ test_that("jointModel input checks work", {
                                       site.cov = site.cov),
                           cov = c('var_a','var_b'),
                           multicore = FALSE),
-               paste0("Data should include 'qPCR.N', 'qPCR.K', ",
-                      "'count', and 'site.cov'."))
+               paste(paste0("Data should include 'qPCR.N', 'qPCR.K', ",
+                            "'count', and 'site.cov'."),
+                     'See the eDNAjoint guide for data formatting help: ',
+                     paste0('https://bookdown.org/abigailkeller/eDNAjoint_',
+                            'vignette/usecase2.html#prepare-the-data'),
+                     sep='\n'))
   #3. input tags are valid, q = TRUE, cov = FALSE
   expect_error(jointModel(data = list(qPCR.k = rbind(c(1,1,1),c(1,1,NA)),
                                       qPCR.n = rbind(c(3,3,3),c(3,3,NA)),
@@ -25,8 +33,12 @@ test_that("jointModel input checks work", {
                                       count.type = rbind(c(1,2,1),c(1,2,NA))),
                           q = TRUE,
                           multicore = FALSE),
-               paste0("Data should include 'qPCR.N', 'qPCR.K', ",
-                      "'count', and 'count.type'."))
+               paste(paste0("Data should include 'qPCR.N', 'qPCR.K', ",
+                            "'count', and 'count.type'."),
+                     'See the eDNAjoint guide for data formatting help: ',
+                     paste0('https://bookdown.org/abigailkeller/eDNAjoint_',
+                            'vignette/usecase3.html#prepare-the-data'),
+                     sep='\n'))
   #4. input tags are valid, q = TRUE, cov = TRUE
   site.cov <- cbind(c(1,0),c(0.4,-0.4))
   colnames(site.cov) <- c('var_a','var_b')
@@ -37,8 +49,12 @@ test_that("jointModel input checks work", {
                                       site.cov = site.cov),
                           cov = c('var_a','var_b'),q = TRUE,
                           multicore = FALSE),
-               paste0("Data should include 'qPCR.N', 'qPCR.K', 'count', ",
-                      "'count.type', and 'site.cov'."))
+               paste(paste0("Data should include 'qPCR.N', 'qPCR.K', 'count', ",
+                            "'count.type', and 'site.cov'."),
+                     'See the eDNAjoint guide for data formatting help: ',
+                     paste0('https://bookdown.org/abigailkeller/eDNAjoint_',
+                            'vignette/usecase3.html#prepare-the-data'),
+                     sep='\n'))
 
   #5. make sure dimensions of qPCR.N and qPCR.K are equal
   #' @srrstats {BS2.1a} Test to ensure pre-processing routines to ensure all
@@ -47,7 +63,11 @@ test_that("jointModel input checks work", {
[…]
