+++ R/jointModel.R
@@ -439,6 +439,23 @@ jointModel <- function(data, cov = NULL, family = 'poisson',
   stopifnot(is.double(sum(colMeans(rstan::extract(out,
                                                   par = 'log_lik')$log_lik))))
 
+  # check for divergent transitions
+  div_trans <- sum(lapply(rstan::get_sampler_params(modelfit,
+                                                    inc_warmup = FALSE),
+                          div_check)[[1]])
+  # print either troubleshooting or visualization tips
+  if(div_trans>0){
+    url <- paste0('https://bookdown.org/abigailkeller/eDNAjoint_vignette/',
+                  'tips.html#visualization-tips')
+    message <- 'Refer to the eDNAjoint guide for troubleshooting tips: '
+    cat(message, url, "\n")
+  } else {
+    url <- paste0('https://bookdown.org/abigailkeller/eDNAjoint_vignette/',
+                  'tips.html#visualization-tips')
+    message <- 'Refer to the eDNAjoint guide for visualization tips: '
+    cat(message, url, "\n")
+  }
+
   # add chain names to init list
   names(inits) <- paste0('chain',seq(1,n.chain,1))
 
@@ -724,15 +741,23 @@ catchability_checks <- function(data,cov){
   #cov='None'
   if (all(is.null(cov)) && !all(c('qPCR.N', 'qPCR.K',
                                   'count','count.type') %in% names(data))){
-    errMsg <- paste0("Data should include 'qPCR.N', 'qPCR.K', ",
+    errMsg1 <- paste0("Data should include 'qPCR.N', 'qPCR.K', ",
                      "'count', and 'count.type'.")
+    errMsg2 <- 'See the eDNAjoint guide for data formatting help: '
+    errMsg3 <- paste0('https://bookdown.org/abigailkeller/eDNAjoint_vignette',
+                      '/usecase3.html#prepare-the-data')
+    errMsg <- paste(errMsg1,errMsg2,errMsg3, sep = "\n")
     stop(errMsg)
   }
   #q=TRUE and cov != 'None'
   if (all(!is.null(cov)) && !all(c('qPCR.N', 'qPCR.K', 'count',
                                    'count.type','site.cov') %in% names(data))){
-    errMsg <- paste0("Data should include 'qPCR.N', 'qPCR.K', ",
+    errMsg1 <- paste0("Data should include 'qPCR.N', 'qPCR.K', ",
                      "'count', 'count.type', and 'site.cov'.")
+    errMsg2 <- 'See the eDNAjoint guide for data formatting help: '
+    errMsg3 <- paste0('https://bookdown.org/abigailkeller/eDNAjoint_vignette',
+                      '/usecase3.html#prepare-the-data')
+    errMsg <- paste(errMsg1,errMsg2,errMsg3, sep = "\n")
     stop(errMsg)
   }
 
@@ -740,14 +765,22 @@ catchability_checks <- function(data,cov){
   #' @srrstats {G5.8,G5.8a} Pre-processing routines to check for zero-length
   #'   data
   if (dim(data$count.type)[1] == 0) {
-    errMsg <- "count.type contains zero-length data."
+    errMsg1 <- "count.type contains zero-length data."
+    errMsg2 <- 'See the eDNAjoint guide for data formatting help: '
+    errMsg3 <- paste0('https://bookdown.org/abigailkeller/eDNAjoint_vignette',
+                      '/usecase3.html#prepare-the-data')
+    errMsg <- paste(errMsg1,errMsg2,errMsg3, sep = "\n")
     stop(errMsg)
   }
[…]
