+++ README.md
@@ -27,6 +27,8 @@ as estimation of catchability coefficients when multiple traditional
 gear types are used. Additional functions in the package facilitate
 interpretation of model fits.
 
+<img src="man/figures/basic_diagram_final-01.png" width="50%" style="display: block; margin: auto;" />
+
 ## Installation
 
 You can install the development version of eDNAjoint. Note that
@@ -68,7 +70,7 @@ detection, $p_{10}$:
 # summarize p10 posterior
 jointSummarize(goby.fit$model, par = 'p10')
 #>      mean se_mean    sd  2.5% 97.5%    n_eff Rhat
-#> p10 0.003       0 0.001 0.001 0.007 16465.27    1
+#> p10 0.003       0 0.001 0.001 0.007 16195.17    1
 ```
 
 Or to find the number of eDNA samples and traditional survey samples
