### 2024-08-14 #642 @abigailkeller
@smwindecker and @chitrams thank you so much again for your thorough review! I really appreciate it. And thanks @emitanaka for facilitating.

Below are the reviewer comments in bold with my responses beneath. Please let me know if there is anything that can still be improved:

**G1.2: The reason this was erroring was because the function is looking for "list(markdown = TRUE)" in the description file. The additional content you have ", roclets = c ("namespace", "rd", "srr::srr_stats_roclet"))" is making the uses_markdown_md() return an NA value and thereby erroring the use_lifecycle() function, as described in https://rdrr.io/cran/usethis/src/R/roxygen.R#sym-uses_roxygen_md. If you temporarily remove the roclets argument, you can run the usethis::use_lifecycle() then paste back the roclets argument into the description file afterwards.**

Thanks - I don't think I would have figured this out myself! My [README file](https://github.com/abigailkeller/eDNAjoint/blob/master/README.md) now has a lifecycle badge for the package.

**BS6.0/BS6.1 re default print and plot methods. We agree with the authors that the existing packages that exist to plot stan outputs are sufficient and they do not need to provide their own. I also think the demonstrations and examples in the bookdown are very helpful. We also agree that the rstan.fit object has its own print function, however the jointModel output contains not only the rstan.fit object but also a list of the inits. If they would like, the authors could perhaps try to return the inits in a neater format, for the moment they are returned in an unnamed list of 4, which, to a novice user, may not be immediately obvious correpond to the 4 chains.**

Good point! I have updated the code so that the `jointModel` function returns a named list (see [lines 459-460 in jointModel.R](https://github.com/abigailkeller/eDNAjoint/blob/master/R/jointModel.R)).

**Some guidelines for contributing could be listed explicitly if the authors would welcome them.**

I have added both [contributing guidelines](https://github.com/abigailkeller/eDNAjoint/blob/master/CONTRIBUTING.md) and a [code of conduct](https://github.com/abigailkeller/eDNAjoint/blob/master/CODE_OF_CONDUCT.md) to the package repo.

**In general the existing visualisations in other packages for stan model outputs are sufficient for visualisation of the outputs from this package. The guidelines in the bookdown are also very clear. However for users that don't work through the bookdown itself, if the authors were interested, they could perhaps point to these options in the help file for the model fit function or in a print message for the fitted object. I spoke to a few colleagues who use similar packages that wrap stan and they agreed they wouldn't expect bespoke plotting functionality.**

I like it! I've added a print message in [lines 453-457 of jointModel.R](https://github.com/abigailkeller/eDNAjoint/blob/master/R/jointModel.R) to link users to the [Visualization Tips](https://bookdown.org/abigailkeller/eDNAjoint_vignette/tips.html#visualization-tips) section of the user guide.

**As an aside, after speaking with a colleague who works in eDNA, they expressed interest in the intent of the package but noted that non perfectly paired data is more common. It would be great in the future if the authors could consider an extension that would allow users to input non-paired data, to allow more potential users to get the most out of available datasets (though I understand this is a much thornier model, just thought I'd pass along the note).**

I really like this suggestion, thanks! I've updated [the models](https://github.com/abigailkeller/eDNAjoint/tree/master/inst/stan) to accommodate what I am calling "semi-paired data". Instead of requiring the user to have both paired eDNA and traditional survey data at all sites, now the user can have paired data at some sites and eDNA data only at other sites. The models now use the estimated relationship between eDNA and traditional survey data from paired sites to make predictions at the unpaired sites. I have also added [another use case](https://bookdown.org/abigailkeller/eDNAjoint_vignette/usecase4.html) to the user guide to demonstrate this new functionality.

However, I imagine your colleague may also be referring to "non perfectly paired" data in the sense that the paired data do no align perfectly in space and/or time. I'm definitely interested in addressing this in a future extension. :)

**Should the author(s) deem it appropriate, we agree to be acknowledged as a package reviewer ("rev" role) in the package DESCRIPTION file.**

Yes! I'd be happy to add you both. @smwindecker and @chitrams what are your email addresses?

**We had one issue with installation, which was that for one of us the install took around 30 minutes to complete. This seems unusally long. If this install time cannot be reduced, perhaps just a note in the readme warning users that it may take a long time. The clang messages that come along with the compiled stan models are confronting for a user that has not encountered them before, and may make someone interrupt installation if it is taking so long.**

After the package is through review, I am intending to put the package on Cran and give Cran installation instructions in the repo README. Cran provides pre-compiled binary files, while installing from source requires compiling the models, which I imagine is why it took so long to install and was giving stressful messages! I am hoping installation will be smoother from Cran, but I also have [added more detail to the README](https://github.com/abigailkeller/eDNAjoint/blob/master/README.md) that users may expect long installation times from source.

**A few notes from some automatic checks:**

**The following function have high cyclomatic complexity (>50): detectionCalculate (117), detectionPlot (117), jointSummarize (105), jointModel (66). You can make them easier to reason about by encapsulating distinct steps of your function into subfunctions.**

Thanks! I have made the code more modular and vectorized to reduce cyclomatic complexity. I believe the cyclomatic complexity of detectionCalculate, detectionPlot, jointsSummarize, and jointModel is now 13, 3, 4, and 20, respectively.

**use '<-' for assignment instead of '='. '<-' is the standard, and R users and developers are used it and it is easier to read your code for them if you use '<-'. See in tests/testthat/test-jointSelect.R:35:10 tests/testthat/test-jointSelect.R:38:10 tests/testthat/test-muCritical.R:57:11 tests/testthat/test-muCritical.R:63:9**

I have updated the assignment syntax in [test-muCritical.R](https://github.com/abigailkeller/eDNAjoint/blob/master/tests/testthat/test-muCritical.R) and [test-jointSelect.R](https://github.com/abigailkeller/eDNAjoint/blob/master/tests/testthat/test-jointSelect.R).

**style consistency: add spaces around "=" for better readability.**

I have made sure there are spaces around "=" in all files of the package.
