### 2024-08-14 223560f
fixed print statement in jointModel
