### 2024-08-20 3deafbc
removed detectionPlot title
