### 2024-08-22 d0e8f9a
updated labels in detectionPlot
