### 2024-09-05 e3d9752
fixed detectioncalculate function description
