+++ src/stanExports_traditional_catchability_gamma.h
@@ -29,54 +29,54 @@ using namespace stan::math;
 stan::math::profile_map profiles__;
 static constexpr std::array<const char*, 49> locations_array__ =
   {" (found before start of program)",
-  " (in 'string', line 10, column 4 to column 36)",
-  " (in 'string', line 11, column 4 to column 39)",
-  " (in 'string', line 12, column 4 to column 44)",
-  " (in 'string', line 15, column 4 to column 28)",
-  " (in 'string', line 16, column 4 to column 35)",
-  " (in 'string', line 32, column 2 to column 20)",
-  " (in 'string', line 33, column 2 to column 20)",
-  " (in 'string', line 34, column 2 to column 28)",
-  " (in 'string', line 18, column 6 to column 48)",
-  " (in 'string', line 17, column 17 to line 19, column 5)",
-  " (in 'string', line 17, column 4 to line 19, column 5)",
-  " (in 'string', line 21, column 6 to column 42)",
-  " (in 'string', line 20, column 17 to line 22, column 5)",
-  " (in 'string', line 20, column 4 to line 22, column 5)",
-  " (in 'string', line 35, column 2 to column 18)",
-  " (in 'string', line 37, column 4 to column 31)",
-  " (in 'string', line 36, column 18 to line 38, column 3)",
-  " (in 'string', line 36, column 2 to line 38, column 3)",
-  " (in 'string', line 40, column 4 to column 38)",
-  " (in 'string', line 39, column 21 to line 41, column 3)",
-  " (in 'string', line 39, column 2 to line 41, column 3)",
-  " (in 'string', line 43, column 10 to column 80)",
-  " (in 'string', line 42, column 17 to line 44, column 7)",
-  " (in 'string', line 42, column 4 to line 44, column 7)",
-  " (in 'string', line 26, column 6 to column 57)",
-  " (in 'string', line 25, column 17 to line 27, column 5)",
-  " (in 'string', line 25, column 4 to line 27, column 5)",
-  " (in 'string', line 28, column 4 to column 28)",
-  " (in 'string', line 29, column 4 to column 29)",
-  " (in 'string', line 2, column 4 to column 19)",
-  " (in 'string', line 3, column 10 to column 11)",
-  " (in 'string', line 3, column 4 to column 28)",
-  " (in 'string', line 4, column 4 to column 22)",
-  " (in 'string', line 5, column 10 to column 11)",
-  " (in 'string', line 5, column 4 to column 29)",
-  " (in 'string', line 6, column 4 to column 25)",
-  " (in 'string', line 7, column 11 to column 12)",
-  " (in 'string', line 7, column 13 to column 20)",
-  " (in 'string', line 7, column 4 to column 26)",
-  " (in 'string', line 10, column 10 to column 14)",
-  " (in 'string', line 11, column 10 to column 14)",
-  " (in 'string', line 12, column 27 to column 34)",
-  " (in 'string', line 15, column 20 to column 21)",
-  " (in 'string', line 16, column 10 to column 11)",
-  " (in 'string', line 32, column 9 to column 16)",
-  " (in 'string', line 33, column 9 to column 10)",
-  " (in 'string', line 34, column 9 to column 13)",
-  " (in 'string', line 34, column 14 to column 23)"};
+  " (in 'traditional_catchability_gamma', line 10, column 4 to column 36)",
+  " (in 'traditional_catchability_gamma', line 11, column 4 to column 39)",
+  " (in 'traditional_catchability_gamma', line 12, column 4 to column 44)",
+  " (in 'traditional_catchability_gamma', line 15, column 4 to column 28)",
+  " (in 'traditional_catchability_gamma', line 16, column 4 to column 35)",
[…]
