+++ src/stanExports_traditional_gamma.h
@@ -29,36 +29,36 @@ using namespace stan::math;
 stan::math::profile_map profiles__;
 static constexpr std::array<const char*, 31> locations_array__ =
   {" (found before start of program)",
-  " (in 'string', line 8, column 4 to column 36)",
-  " (in 'string', line 9, column 4 to column 38)",
-  " (in 'string', line 12, column 4 to column 35)",
-  " (in 'string', line 25, column 2 to column 20)",
-  " (in 'string', line 26, column 2 to column 18)",
-  " (in 'string', line 14, column 6 to column 42)",
-  " (in 'string', line 13, column 17 to line 15, column 5)",
-  " (in 'string', line 13, column 4 to line 15, column 5)",
-  " (in 'string', line 28, column 4 to column 29)",
-  " (in 'string', line 27, column 18 to line 29, column 3)",
-  " (in 'string', line 27, column 2 to line 29, column 3)",
-  " (in 'string', line 31, column 10 to column 72)",
-  " (in 'string', line 30, column 17 to line 32, column 7)",
-  " (in 'string', line 30, column 4 to line 32, column 7)",
-  " (in 'string', line 19, column 6 to column 49)",
-  " (in 'string', line 18, column 17 to line 20, column 5)",
-  " (in 'string', line 18, column 4 to line 20, column 5)",
-  " (in 'string', line 21, column 3 to column 27)",
-  " (in 'string', line 22, column 3 to column 28)",
-  " (in 'string', line 2, column 4 to column 19)",
-  " (in 'string', line 3, column 10 to column 11)",
-  " (in 'string', line 3, column 4 to column 28)",
-  " (in 'string', line 4, column 4 to column 22)",
-  " (in 'string', line 5, column 10 to column 11)",
-  " (in 'string', line 5, column 4 to column 29)",
-  " (in 'string', line 8, column 10 to column 14)",
-  " (in 'string', line 9, column 10 to column 14)",
-  " (in 'string', line 12, column 10 to column 11)",
-  " (in 'string', line 25, column 9 to column 10)",
-  " (in 'string', line 26, column 9 to column 13)"};
+  " (in 'traditional_gamma', line 8, column 4 to column 36)",
+  " (in 'traditional_gamma', line 9, column 4 to column 38)",
+  " (in 'traditional_gamma', line 12, column 4 to column 35)",
+  " (in 'traditional_gamma', line 25, column 2 to column 20)",
+  " (in 'traditional_gamma', line 26, column 2 to column 18)",
+  " (in 'traditional_gamma', line 14, column 6 to column 42)",
+  " (in 'traditional_gamma', line 13, column 17 to line 15, column 5)",
+  " (in 'traditional_gamma', line 13, column 4 to line 15, column 5)",
+  " (in 'traditional_gamma', line 28, column 4 to column 29)",
+  " (in 'traditional_gamma', line 27, column 18 to line 29, column 3)",
+  " (in 'traditional_gamma', line 27, column 2 to line 29, column 3)",
+  " (in 'traditional_gamma', line 31, column 10 to column 72)",
+  " (in 'traditional_gamma', line 30, column 17 to line 32, column 7)",
+  " (in 'traditional_gamma', line 30, column 4 to line 32, column 7)",
+  " (in 'traditional_gamma', line 19, column 6 to column 49)",
+  " (in 'traditional_gamma', line 18, column 17 to line 20, column 5)",
+  " (in 'traditional_gamma', line 18, column 4 to line 20, column 5)",
+  " (in 'traditional_gamma', line 21, column 3 to column 27)",
+  " (in 'traditional_gamma', line 22, column 3 to column 28)",
[…]
