### 2024-09-13 fc577af
fixed initial value problem
