+++ R/jointModel.R
@@ -340,7 +340,7 @@ jointModel <- function(data, cov = NULL, family = 'poisson',
   sigma_ln <- sqrt(sigma_2_ln)
 
   # create data that will be present in all model variations
-  data <- list(
+  model_data <- list(
     S = nrow(qPCR_all_trad),
     S_dna = nrow(qPCR_all_dna),
     C = nrow(count_all),
@@ -362,23 +362,23 @@ jointModel <- function(data, cov = NULL, family = 'poisson',
 
   # append data if q == TRUE
   if(isCatch_type(q)){
-    data <- rlist::list.append(
-      data,
+    model_data <- rlist::list.append(
+      model_data,
       nparams = length(q_names),
       mat = as.matrix(count_all[,q_names])
     )
   }
   # append data if family == negbin
   if(isNegbin_type(family)){
-    data <- rlist::list.append(
-      data,
+    model_data <- rlist::list.append(
+      model_data,
       phipriors = phipriors
     )
   }
   # append data if cov != NULL
   if(isCov_type(cov)){
-    data <- rlist::list.append(
-      data,
+    model_data <- rlist::list.append(
+      model_data,
       nsitecov = length(cov)+1,
       mat_site = as.matrix(site_mat)
     )
@@ -399,9 +399,9 @@ jointModel <- function(data, cov = NULL, family = 'poisson',
 
   # get initial values
   if(isCatch_type(q)){
-    inits <- get_inits(n.chain,qPCR_all,initial_values,cov,q_names)
+    inits <- get_inits(n.chain,qPCR_all,initial_values,cov,q_names,data)
   } else {
-    inits <- get_inits(n.chain,qPCR_all,initial_values,cov)
+    inits <- get_inits(n.chain,qPCR_all,initial_values,cov,data)
   }
 
   # run model
@@ -418,7 +418,7 @@ jointModel <- function(data, cov = NULL, family = 'poisson',
       stanmodels$joint_binary_pois,
       stanmodels$joint_binary_negbin,
       stanmodels$joint_binary_gamma)[model_index][[1]],
-    data = data,
+    data = model_data,
     cores = cores,
     seed = SEED,
     #' @srrstats {G2.4,G2.4a} explicit conversion to
@@ -528,26 +528,26 @@ get_stan_model <- function(q, family, cov){
 #'   chain
 
 
-get_inits <- function(n.chain,qPCR_all,initial_values,cov,q_names=NULL){
+get_inits <- function(n.chain,qPCR_all,initial_values,cov,data,q_names=NULL){
   if(!is.null(cov)){
     if(!is.null(q_names)){
       inits <- init_joint_cov_catchability(n.chain,qPCR_all,q_names,cov,
-                                           initial_values)
+                                           initial_values,data)
     } else {
-      inits <- init_joint_cov(n.chain,qPCR_all,cov,initial_values)
+      inits <- init_joint_cov(n.chain,qPCR_all,cov,initial_values,data)
     }
   } else {
     if(!is.null(q_names)){
       inits <- init_joint_catchability(n.chain,qPCR_all,q_names,
-                                       initial_values)
+                                       initial_values,data)
     } else {
-      inits <- init_joint(n.chain,qPCR_all,initial_values)
+      inits <- init_joint(n.chain,qPCR_all,initial_values,data)
     }
   }
   return(inits)
 }
 
-init_joint_cov <- function(n.chain,qPCR_all,cov,initial_values){
+init_joint_cov <- function(n.chain,qPCR_all,cov,initial_values,data){
   # helper function
   # joint model, catchability coefficient, site covariates
   A <- list()
[…]
