### 2024-09-13 109099d
updated initial value code
