+++ tests/testthat/test-jointModel.R
@@ -457,7 +457,7 @@ test_that("jointModel input checks work", {
   for(i in 1:n.chain){
     inits[[i]] <- list(
       mu <- stats::runif(3, 0.01, 5),
-      p10 <- stats::runif(1,log(0.0001),log(0.08)),
+      p10 <- exp(stats::runif(1,log(0.0001),log(0.08))),
       alpha <- rep(0.1,3)
     )
   }
@@ -485,7 +485,7 @@ test_that("jointModel input checks work", {
   for(i in 1:n.chain){
     inits[[i]] <- list(
       mu <- stats::runif(3, -1, 0),
-      p10 <- stats::runif(1,log(0.0001),log(0.08)),
+      p10 <- exp(stats::runif(1,log(0.0001),log(0.08))),
       alpha <- rep(0.1,3)
     )
     names(inits[[i]]) <- c('mu','p10','alpha')
@@ -511,7 +511,7 @@ test_that("jointModel input checks work", {
   for(i in 1:n.chain){
     inits[[i]] <- list(
       mu <- stats::runif(4, 0.1, 1),
-      p10 <- stats::runif(1,log(0.0001),log(0.08)),
+      p10 <- exp(stats::runif(1,log(0.0001),log(0.08))),
       alpha <- rep(0.1,3)
     )
     names(inits[[i]]) <- c('mu','p10','alpha')
@@ -565,7 +565,7 @@ test_that("jointModel input checks work", {
   for(i in 1:n.chain){
     inits[[i]] <- list(
       mu <- stats::runif(2,0,1),
-      p10 <- stats::runif(2,log(0.0001),log(0.08)),
+      p10 <- exp(stats::runif(2,log(0.0001),log(0.08))),
       alpha <- rep(0.1,3)
     )
     names(inits[[i]]) <- c('mu','p10','alpha')
@@ -591,7 +591,7 @@ test_that("jointModel input checks work", {
   for(i in 1:n.chain){
     inits[[i]] <- list(
       mu <- stats::runif(2,0,1),
-      p10 <- stats::runif(1,log(0.0001),log(0.08)),
+      p10 <- exp(stats::runif(1,log(0.0001),log(0.08))),
       beta <- "1"
     )
     names(inits[[i]]) <- c('mu','p10','beta')
@@ -614,7 +614,7 @@ test_that("jointModel input checks work", {
   for(i in 1:n.chain){
     inits[[i]] <- list(
       mu <- stats::runif(2,0,1),
-      p10 <- stats::runif(1,log(0.0001),log(0.08)),
+      p10 <- exp(stats::runif(1,log(0.0001),log(0.08))),
       beta <- c(1,0)
     )
     names(inits[[i]]) <- c('mu','p10','beta')
@@ -637,7 +637,7 @@ test_that("jointModel input checks work", {
   for(i in 1:n.chain){
     inits[[i]] <- list(
       mu <- stats::runif(2,0,1),
-      p10 <- stats::runif(1,log(0.0001),log(0.08)),
+      p10 <- exp(stats::runif(1,log(0.0001),log(0.08))),
       alpha <- c("1","2")
     )
     names(inits[[i]]) <- c('mu','p10','alpha')
@@ -663,7 +663,7 @@ test_that("jointModel input checks work", {
   for(i in 1:n.chain){
     inits[[i]] <- list(
       mu <- stats::runif(2,0,1),
-      p10 <- stats::runif(1,log(0.0001),log(0.08)),
+      p10 <- exp(stats::runif(1,log(0.0001),log(0.08))),
       alpha <- rep(0.1,2)
     )
     names(inits[[i]]) <- c('mu','p10','alpha')
[…]
