+++ R/jointModel.R
@@ -399,9 +399,11 @@ jointModel <- function(data, cov = NULL, family = 'poisson',
 
   # get initial values
   if(isCatch_type(q)){
-    inits <- get_inits(n.chain,qPCR_all,initial_values,cov,data,q_names)
+    inits <- get_inits(n.chain,qPCR_all,initial_values,cov,L_match_trad,
+                       L_match_dna,data,q_names)
   } else {
-    inits <- get_inits(n.chain,qPCR_all,initial_values,cov,data)
+    inits <- get_inits(n.chain,qPCR_all,initial_values,cov,L_match_trad,
+                       L_match_dna,data)
   }
 
   # run model
@@ -528,26 +530,41 @@ get_stan_model <- function(q, family, cov){
 #'   chain
 
 
-get_inits <- function(n.chain,qPCR_all,initial_values,cov,data,q_names=NULL){
+get_inits <- function(n.chain,qPCR_all,initial_values,cov,L_match_trad,
+                      L_match_dna,data,q_names=NULL){
   if(!is.null(cov)){
     if(!is.null(q_names)){
       inits <- init_joint_cov_catchability(n.chain,qPCR_all,q_names,cov,
-                                           initial_values,data)
+                                           initial_values,L_match_trad,
+                                           L_match_dna,data)
     } else {
-      inits <- init_joint_cov(n.chain,qPCR_all,cov,initial_values,data)
+      inits <- init_joint_cov(n.chain,qPCR_all,cov,initial_values,
+                              L_match_trad,L_match_dna,data)
     }
   } else {
     if(!is.null(q_names)){
       inits <- init_joint_catchability(n.chain,qPCR_all,q_names,
-                                       initial_values,data)
+                                       initial_values,L_match_trad,
+                                       L_match_dna,data)
     } else {
-      inits <- init_joint(n.chain,qPCR_all,initial_values,data)
+      inits <- init_joint(n.chain,qPCR_all,initial_values,L_match_trad,
+                          L_match_dna,data)
     }
   }
   return(inits)
 }
 
-init_joint_cov <- function(n.chain,qPCR_all,cov,initial_values,data){
+init_joint_cov <- function(n.chain,qPCR_all,cov,initial_values,
+                           L_match_trad,L_match_dna,data){
+
+  # get mu means
+  mu_means_trad <- as.vector(na.omit(rowMeans(data$count,na.rm=TRUE)))
+  mu_means_all <- rep(NA,dim(L_match_dna)[1]+dim(L_match_trad)[1])
+  mu_means_all[L_match_trad$L] <- mu_means_trad
+  if(dim(L_match_dna)[1]>0){
+    mu_means_all[L_match_dna$L] <- rep(mean(mu_means_trad),dim(L_match_dna)[1])
+  }
+
   # helper function
   # joint model, catchability coefficient, site covariates
   A <- list()
@@ -555,9 +572,15 @@ init_joint_cov <- function(n.chain,qPCR_all,cov,initial_values,data){
     for(i in 1:n.chain){
       A[[i]] <- list(
         if('mu' %in% names(initial_values[[i]])){
-          mu_trad <- initial_values[[i]]$mu
+          mu_trad <- initial_values[[i]]$mu[L_match_trad$L]
         } else {
-          mu_trad <- as.numeric(na.omit(rowMeans(data$count,na.rm=TRUE)))
+          mu_trad <- mu_means_trad
+        },
+
[…]

[…]
