### 2024-09-13 b915303
made inits less fixed
