+++ R/jointModel.R
@@ -592,21 +592,22 @@ init_joint_cov <- function(n.chain,qPCR_all,cov,initial_values,
 
         if('alpha' %in% names(initial_values[[i]])){
           alpha <- initial_values[[i]]$alpha
-        } else {
-          alpha <- rep(0.1,length(cov)+1)
         }
       )
-      names(A[[i]]) <- c('mu_trad','mu','log_p10','alpha')
+      if('alpha' %in% names(initial_values[[i]])){
+        names(A[[i]]) <- c('mu_trad','mu','log_p10','alpha')
+      } else {
+        names(A[[i]]) <- c('mu_trad','mu','log_p10')
+      }
     }
   } else {
     for(i in 1:n.chain){
       A[[i]] <- list(
         mu_trad <- mu_means_trad,
         mu <- mu_means_all,
-        log_p10 <- stats::runif(1,log(0.0001),log(0.08)),
-        alpha <- rep(0.1,length(cov)+1)
+        log_p10 <- stats::runif(1,log(0.0001),log(0.08))
       )
-      names(A[[i]]) <- c('mu_trad','mu','log_p10','alpha')
+      names(A[[i]]) <- c('mu_trad','mu','log_p10')
     }
   }
 
@@ -643,12 +644,6 @@ init_joint_cov_catchability <- function(n.chain,qPCR_all,q_names,cov,
           mu <- mu_means_all
         },
 
-        if('q' %in% names(initial_values[[i]])){
-          q_trans <- as.data.frame(initial_values[[i]]$q)
-        } else {
-          q_trans <- as.data.frame(stats::runif(length(q_names),0.01,1))
-        },
-
         if('p10' %in% names(initial_values[[i]])){
           log_p10 <- log(initial_values[[i]]$p10)
         } else {
@@ -657,22 +652,35 @@ init_joint_cov_catchability <- function(n.chain,qPCR_all,q_names,cov,
 
         if('alpha' %in% names(initial_values[[i]])){
           alpha <- initial_values[[i]]$alpha
-        } else {
-          alpha <- rep(0.1,length(cov)+1)
+        },
+
+        if('q' %in% names(initial_values[[i]])){
+          q_trans <- as.data.frame(initial_values[[i]]$q)
         }
       )
-      names(A[[i]]) <- c('mu_trad_1','mu','q_trans','log_p10','alpha')
+      if('q' %in% names(initial_values[[i]])){
+        if('alpha' %in% names(initial_values[[i]])){
+          names(A[[i]]) <- c('mu_trad_1','mu','log_p10','alpha','q_trans')
+        }
+        else {
+          names(A[[i]]) <- c('mu_trad_1','mu','log_p10','q_trans')
+        }
+      } else {
+        if('alpha' %in% names(initial_values[[i]])){
+          names(A[[i]]) <- c('mu_trad_1','mu','log_p10','alpha')
+        } else {
+          names(A[[i]]) <- c('mu_trad_1','mu','log_p10')
+        }
+      }
     }
   } else {
     for(i in 1:n.chain){
       A[[i]] <- list(
         mu_trad_1 <- mu_means_trad,
         mu <- mu_means_all,
-        q_trans <- as.data.frame(stats::runif(length(q_names),0.01,1)),
-        log_p10 <- stats::runif(1,log(0.0001),log(0.08)),
-        alpha <- rep(0.1,length(cov)+1)
+        log_p10 <- stats::runif(1,log(0.0001),log(0.08))
       )
-      names(A[[i]]) <- c('mu_trad_1','mu','q_trans','log_p10','alpha')
+      names(A[[i]]) <- c('mu_trad_1','mu','log_p10')
     }
   }
 
[…]
