### 2024-09-20 6345833
added more structure to inits
