### 2024-09-21 cc3a8aa
fixed alpha inits
