### 2024-09-21 0402e42
changed p10 and alpha/beta inits
