### 2024-09-25 #642 @emitanaka
Thank you for your patience @abigailkeller (and apologies, I'm a juggling hell).

@smwindecker and @chitrams, thank you so much for your thorough review! Could I ask you to complete the review with the template [here](https://devguide.ropensci.org/approval2template.html) to indicate your approval?

@mpadge I have tried to source other reviewers but I hadn't had luck. Given Saras and Chitra have done a thorough review (and I've gone through the responses myself), would that suffice from the rOpenSci perpsective?
