### 2024-09-25 #642 @mpadge
@emitanaka The joint review is absolutely fine. In order for our bot to separately log these contributions, we'll nevertheless need two separate review statements.

@chitrams: Could you please make a comment here that your review is included in [the previous joint review](https://github.com/ropensci/software-review/issues/642#issuecomment-2238886129), include a link to that review, and your estimated review duration? Note that you should do this prior to Emi's request for you to complete the approval template. Thanks!

@emitanaka You'll then need to officially log the reviews, as described [in our *Dev Guide*](https://github.com/ropensci/software-review/issues/642#issuecomment-2238886129) by issuing `@ropensci-review-bot submit review <link to review comment> time <time in hours>`. The `<link to review comment>` will need to link to the main review to log @smwindecker, and to the separate comment of @chitrams to log that review.

After that, things should all be good for the final approval steps. Nice work all! :rocket:
