### 2024-09-26 #642 @smwindecker
## Reviewer Response

#### Final approval (post-review)

Estimated hours spent reviewing: Review 1, 9.5 hours, Reviewer 2, 12.5 hours
