### 2024-09-26 #642 @chitrams
Hello! My review is included in the [previous joint review](https://github.com/ropensci/software-review/issues/642#issuecomment-2238886129) and the estimated review duration was 12.5 hours for me.
