### 2024-09-26 #642 @ropensci-review-bot
Approved! Thanks @abigailkeller for submitting and @chitrams, @smwindecker for your reviews! :grin:
