### 2024-10-03 #642 @abigailkeller
Thank you so much @mpadge @chitrams @smwindecker @emitanaka for all the work you've done and advice you've given me! This process has been extremely rewarding for me - I've learned so much!
