### 2024-10-03 #642 @abigailkeller
@ropensci/blog-editors I'm interested in writing a blog post about the origin of the package and my experiences developing it. How should I get started?
