### 2025-03-30 #642 @abigailkeller
Hi @steffilazerte!

Thanks for your help with this. Sorry for the delay - I was waiting for a manuscript about the package to be published!

A created a [pull request](https://github.com/ropensci/roweb3/pull/1004) with a blog post. However, it looks like the draft has some errors associated with Netlify rendering. Do you have any tips?

Best,
Abby
